+++ eth_defi/erc_4626/core.py
@@ -395,6 +395,12 @@ class ERC4626Feature(enum.Enum):
     #: https://yield.fi/
     yieldfi_like = "yieldfi_like"
 
+    #: Resolv
+    #:
+    #: Delta-neutral stablecoin protocol with wstUSR wrapped staking vault.
+    #: https://resolv.xyz/
+    resolv_like = "resolv_like"
+
 
 def get_vault_protocol_name(features: set[ERC4626Feature]) -> str:
     """Deduct vault protocol name based on Vault smart contract features.
@@ -567,6 +573,9 @@ def get_vault_protocol_name(features: set[ERC4626Feature]) -> str:
 
     elif ERC4626Feature.yieldfi_like in features:
         return "YieldFi"
+
+    elif ERC4626Feature.resolv_like in features:
+        return "Resolv"
 
     # No idea
     if ERC4626Feature.erc_7540_like in features: