+++ eth_defi/vault/risk.py
@@ -109,6 +109,7 @@ VAULT_PROTOCOL_RISK_MATRIX = {
     "Gearbox": VaultTechnicalRisk.low,
     "Mainstreet Finance": None,
     "YieldFi": None,
+    "Resolv": None,
 }
 
 #: Particular vaults that are broken, misleading or otherwise problematic.